JCF: Issue #534: deal with loopbacks in application plots
The problem this commit deals with is basically, if you have a module
which subscribes on the network to a wildcarded message (say "msg_.*")
and another module in the same application which sends a version of
message that matches the wildcard (say "msg_A") out onto the network,
then how do you plot that?

Previously, when plotting, the logic in the plotter didn't care about
whether a message was sent onto the network. So it was drawing a
connection between the modules making it look as if they communicated
via a queue. With this commit, if we're plotting an application, a
message gets correctly represented as being sent out over the network,
and as being received from the network (with the usual "X" and "O" to
represent "region outside of this diagram").